[level 1] Title: 두 정수 사이의 합, Time: 0.13 ms, Memory: 72.1 MB -BaekjoonHub

+++ 프로그래머스/1/12912. 두 정수 사이의 합/두 정수 사이의 합.java
@@ -9,7 +9,7 @@ class Solution {
         }
         
         for (int i = a; i <= b; i++) {
-            answer += (long) i;
+            answer += i;
         }
         
         return answer;
